perf: General code restructure to not use more than needed memory
